Allow native.existing_rule/s() with --incompatible_existing_rules_immutable_view to be encoded as json and passed as **kwargs
To do so, we make ExistingRuleView and ExistingRulesView implement Map. This
requires getting rid of NotRepresentableException in starlarkifyValue (see
discussion in #13829) so that the view object's accessors don't throw. And
that is, arguably, the correct thing to do: the distinction between returning
null for some non-representable attributes vs. throwing an exception for others
was entirely artificial.

This in turn implies a small change in behavior: now, `existing_rule(x).keys()`
will omit all attributes whose values cannot be represented in Starlark. (The
old behavior was to allow some such attributes, and throw an exception when
their value was accessed.)

Fixes https://github.com/bazelbuild/bazel/issues/13829
Fixes https://github.com/bazelbuild/bazel/issues/16256

PiperOrigin-RevId: 478511532
Change-Id: If0c580a4e4eee1739136ea30eddce9eca71995e3