BuildEventServiceModule no longer blocks indefinitely while OOM-crashing.
While OOM-crashing, we can't always cleanly finish uploading the final BEP
events (that will record the crash stack trace). Specifically, if the JVM
is GC-thrashing and cannot make enough progress on mutator threads to
finish writing the BEP to its transports, we must not wait until completion.

We now wait for --bes_oom_finish_upload_timeout before giving up if (and only
if) Blaze is crashing due to an OOM.

PiperOrigin-RevId: 478600636
Change-Id: I92dbcd5eff16d0b6706ff51de25e37b615516131