fix: expand_location binary target inputs
The previous behaviour was that location expansion on native rules would
accept binary targets as inputs and resolve to a single location. This
was not an option on custom rules. This patch aligns the behaviour of
Starlark's ctx.expand_location to the native rules. E.g. a py_binary can
be expanded using location/execpath/rootpath instead of previously only
locations/execpaths/rootpaths.

Signed-off-by: Thulio Ferraz Assis <3149049+f0rmiga@users.noreply.github.com>