Move macOS framework name linking into framework feature
This is a subtle change that moves the `-framework Foo` arguments to the linker after the `-filelist` argument. This matters in the case you have multiple symbols, one that is in your code, where the objects / archives are in the filelist, and one is in one of your pre-compiled framework dependencies. Previously the framework definition would have been preferred. This is inline with the order Xcode uses for this. The difference has cause real issues with prebuilt Swift https://github.com/apple/swift/issues/61287

Closes #16342.

PiperOrigin-RevId: 480496762
Change-Id: Id5c7c839ba7b0e4323090810b8195ce6ef099eb7