Bzlmod: Tweak mappings of WORKSPACE repos when Bzlmod is enabled
What we used to do is to let WORKSPACE repos see all repos backed by Bzlmod modules. This is sometimes helpful but not reliably so, especially since module names ("rules_scala") often differ from the widely accepted repo names ("build_bazel_rules_scala").

This CL changes it so that we let WORKSPACE repos see all repos visible to the root module instead. This makes migration failures more tractable to fix -- if something in WORKSPACE is missing a dependency with a certain name, one could simply add that dependency to the MODULE.bazel file.

PiperOrigin-RevId: 480887708
Change-Id: Ie27b0ab93b010d18de58dee811dbd28b4fad5a2a