Revert to using `matches` insted of `find` with `remote_download_regex`
It is much easier to accidentally write regexes with pathological
performance with `find` than with `matches`. If needed, the `find`
functionality can always be obtained with `matches` by prepending and
appending `.*` as needed.

In addition, common usage scenarios such as matching by file extension
are easier to get right: With `matches`, `jar` will visibly fail to
have an effect and is easily corrected to `.*jar` (or even `.*\.jar`),
whereas with `find` it will silently fetch entire directories that
contain the substring `jar`, potentially causing performance
regressions.