Clarify repo visibility error message
Instead of saying `Repository '@...' is not visible from repository
'@...'`, the message now says `No repository visible as '@...' from
'@...'`.

The previous message made it sound as if there was a repo known as
`@foo` that just happens not to be visible from `@bar` (similar to Java
strict deps or the visibility attribute on rules), when in fact `@foo`
is just a string made up by `@bar` that has no meaning until it is
mentioned in some `use_repo` call.

Also replaces `repository '@@'` with the less cryptic term `main
repository`.