Store start/end timestamps of locally executed actions in the disk cache.
This is necessary to ensure that the generate-xml.sh action associated with
every test, which embeds the test runner wall time, doesn't execute again when
run on a warm disk cache.

This PR makes the following changes:

1. Add SpawnResult#getStartTime alongside existing #getWallTime and populate
   it in various SpawnRunner implementations.
2. Populate the ExecutedActionMetadata with the start/wall times for a
   locally executed spawn when storing it in the disk cache.
3. Populate the SpawnResult with the start/wall times for a spawn
   retrieved from the disk cache.
4. Add a regression test for the scenario described above.

In order to avoid changes to the remote execution protocol, the start/end
time are stored in the existing ActionResult.execution_metadata fields.
Since there is some mismatch between this field and the metrics available
in a SpawnResult, this might be revisited later.

This is a cherry-pick of Bazel upstream commit 93029d8f1f0. It happens
to fix misreporting of test execution durations.

Change-Id: I451fdb0e6ea8451d498e814abf2e236b237b27a2
Reviewed-on: https://gerrit.sg.apple.com/c/swplatform/bazelbuild/bazel/+/380628
Tested-by: Hyperion Jenkins <hyperion_jenkins@apple.com>
Reviewed-by: Mathieu Olivari <molivari@apple.com>