Use the execution time instead of the worker time for the wall time
SpawnResult.getWallTime() is documented as follows:

    Returns the wall time taken by the {@link Spawn}'s execution.

With the math performed as is, we also end up including time spent
downloading inputs and uploading outputs. I don't think that this is in
the spirit of this method. There are also getUserTime() and
getSystemTime(), meaning this was likely modeled after the time(1)
utility.

Change-Id: I385ec6d16b9480462641f3313324a3e0d12cdd8b
Reviewed-on: https://gerrit.sg.apple.com/c/swplatform/bazelbuild/bazel/+/380637
Tested-by: Hyperion Jenkins <hyperion_jenkins@apple.com>
Reviewed-by: Mathieu Olivari <molivari@apple.com>