Use the execution time instead of the worker time for the wall time
SpawnResult.getWallTime() is documented as follows:

    Returns the wall time taken by the {@link Spawn}'s execution.

With the math performed as is, we also end up including time spent
downloading inputs and uploading outputs. I don't think that this is in
the spirit of this method. There are also getUserTime() and
getSystemTime(), meaning this was likely modeled after the time(1)
utility.

Change-Id: Id8683e2109ce2803877eea535da56d13ae77afc7