Clarify repo visibility error message
Instead of saying `Repository '@foo' is not visible from repository '@bar'`, the message now says `No repository visible as '@foo' from '@bar'`.

The previous message made it sound as if there was a repo known as `@foo` that just happens not to be visible from `@bar` (similar to Java strict deps or the visibility attribute on rules), when in fact `@foo` is just a string made up by `@bar` that has no meaning until it is mentioned in some `use_repo` call.

Also replaces `repository '@@'` with the less cryptic term `main repository`.

Closes #16480.

PiperOrigin-RevId: 481596071
Change-Id: I8704255d8bf9c8b6fe725ba6d288d4e8ebdfc8d4