Don't overwrite errors during `bazel run`.
We output the "Build completed successfully" message via the
progress bar (IMHO, as a more long term fix we should fix that). The
progress bar clears previous progress bar contents.
Previously, we did miss to mark the build as no longer running when it
was actually complete. Both at this point of the invocation, and later
when the whole invocation was complete we did output the "Build
completed successfully" message, deleting what we thought would be the
contents of the progress bar, not aware that there was an ERROR message
in between.

The new test illustrates this behavior.

In 7c074b5a5f03a05a108fdfed1910fe75a538a323, `buildComplete` was changed
into `buildRunning` and the boolean value was reversed. But we missed
changing the value on `AfterCommandEvent`.