Split apple starlarkbuildapi into apple and objc
This is needed to break up the circular dependency in some upcoming work, where
some Apple APIs need to depend on cpp.  In general, anything that is
ObjC-related should be moved to objc.  Specifically, this change:

1. moves ObjcProviderApi to objc.
2. moves any API depending on ObjcProviderApi to objc.

This refactoring allows us to add CcInfoApi to objc APIs that carry linking
info, in preparation for migrating linking info to CcInfo.

PiperOrigin-RevId: 481916193
Change-Id: I56f88fd64f617b7c14137e43deb23206eabe9da3