Improve AbruptExitException error messages
When misconfiguring the Remote Output Service (#12823), Bazel may
sometimes throw error messages that look like this:

    ERROR: Couldn't create action output directory

No details about the actual reason for it failing are included. This
change adds some string formatting, so that the error message becomes as
follows:

    ERROR: Couldn't create action output directory: /private/var/tmp/_bazel_ed/.../execroot/.../bazel-out/_tmp/actions (No such file or directory)

This makes it a lot easier to understand what's wrong.