Use `Runtime#availableProcessors` on Linux
As of JDK 11, `Runtime#availableProcessors` on Linux is aware of cgroup
resource limits and thus behaves as expected in containers, whereas
`/proc/cpuinfo` still reports the number of logical processors available
to the host.

Unfortunately, the JVM-internal `os::total_memory` function, which
similarly takes cgroup limits into account when computing the available
RAM, does not seem to be accessible from Java.

https://hg.openjdk.java.net/jdk/hs/rev/7f22774a5f42#l6.178