Experimentally support remote persistent workers.
Add a new `--experimental_remote_mark_tool_inputs` flag, which makes Bazel tag tool inputs when executing actions remotely, and also adds a tools input key to the platform proto sent as part of the remote execution request.

This allows a remote execution system to implement persistent workers, i.e., to keep worker processes around and reuse them for subsequent actions. In a trivial example, this improves build performance by ~3x.

We use "persistentWorkerKey" for the platform property, with the value being a hash of the tool inputs, and "bazel_tool_input" as the node property name, with an empty string as value—this is just a boolean tag.

Fixes https://github.com/bazelbuild/bazel/issues/10091.

Co-authored-by: Ulf Adams <ulf@engflow.com>

Closes #16362.

PiperOrigin-RevId: 482433908
Change-Id: I2a80834731fd0169c08d4beea348f61a323ca028