Add and propagate CcInfo in binary providers that carry linking info
The info in the providers is used to support avoid deps, and we need to add
CcInfo to them in preparation for migrating linking info to CcInfo.

In the starlark constructor of those providers, add an extra cc_info field, and
allow it to be empty while we migrate all the starlark usage.

PiperOrigin-RevId: 482523687
Change-Id: Ibae9c29aadb6819d22d55f207166037352ede0a0