Expand tree artifact headers for include scanning.
When compiling a header module with tree artifact headers, we run includes scanning on the tree itself. This doesn't work as the include scanner is not intended to work on directories. Expand tree artifacts before we pass them to the include scanner.

Base the tree artifact expansion on Skyframe directly since the `ArtifactExpander` only covers direct action inputs.

PiperOrigin-RevId: 482627266
Change-Id: I72ebbdcd47deea6ceb508be6b012353208997b01