Classify the undeclared test outputs directory as a directory instead of a file.
When constructing a remote execution request, any output other than a tree artifact is classified as a (regular) file. But the undeclared outputs directory is actually an ActionInput, and promoting it to an Artifact seems a rather involved refactor. Instead, add an ActionInput#isDirectory method, similar to the preexisting ActionInput#isSymlink, and use it in RemoteExecutionService#buildCommand.

Also take the opportunity to add tests verifying that all output types are handled correctly.

Fixes #16315.

PiperOrigin-RevId: 482738939
Change-Id: I5d237efd7244cee25ed9702b593c6d0c714e3f8a