Check digest of local file to decide whether to re-download it.
Normally, for a non-clean build (outputBase is not clean), skyframe is able to detect modifications, invalid the action, and rerun. Before rerun, it will delete the stales outputs. So we only use `path.exists()` to decide whether we should download an input.

However, there are some files  under the outputBase are not tracked by skyframe. In that case, we could wrongly use a staled output.

PiperOrigin-RevId: 483352318
Change-Id: I7e100100e6c3218630c5dc9bf3f900b2de232e0e