[Skymeld] Sets the Build Configurations in the final result as soon as they're available.
There was a bug which was exposed by the BuildWithoutTheBytesIntegrationTest:

For Skymeld, the BuildConfigurationCollection was only set to the final result
when both analysis & execution finished successfully [1]. In the test, the
execution phase intentionally failed, and trying to access the build
configurations for assertions caused an NPE. This was because the
configurations were never set in the final result: the build did not complete
successfully.

This CL fixes that issue by providing a callback to forward the
BuildConfigurationCollection to the final build result as soon as it's
available.

[1] https://github.com/bazelbuild/bazel/blob/e3e544587fec613fa01e54fb5248fed53c68ba57/src/main/java/com/google/devtools/build/lib/buildtool/BuildTool.java#L331-L340

PiperOrigin-RevId: 483666498
Change-Id: I971c41462c7f408d864394240d807c9416ce44f1