[remote/downloader] Don't include headers in `FetchBlobRequest`
Including the headers in the request is very inefficient as credentials
should never change the content of the downloaded archive. In fact,
given that Bazel verifies the checksum of the downloaded file, the
credentials cannot possibly used in a way where they influence
the outcome of the download (other than deciding whether or not
the user is allowed to download the blob at all). Hence, the
credentials should not be included in the request.

Users that need to send credentials to the remote downloader should
do so by passing the credentials as metadata to the gRPC call.

Note that the remote downloader is behind an experimental flag,
so this change does not need to go thorugh the incompatible change
process.