Improve error on invalid `-//foo` and `-@repo//foo` options
`-//foo` and `-@repo//foo` are always invalid Bazel options, but are
usually meant to be either negative target patterns (which have to come
after the `--` separator) or Starlark options (which always start with
`--`).

With this commit, the error shown to the user mentions these two
situations and how to fix the issue in each case.