Fix import instructions for Python runfiles library
The rules_python version of the runfiles library has diverged and,
since the import statement mentions a repository name, would not work
with Bzlmod.

Instead, use a repo root relative import from `@bazel_tools` and only
switch to rules_python once it has become the source of truth for the
Python rules.