Emit executable target's repository name into repo mapping manifest
Code in the top-level binary target that generates an executable needs
to know its canonical repository name in order to look up runfiles with
Bzlmod. If this information can not be compiled into the binary, e.g.
because it is a shell script and doesn't use a launcher, bootstrapping
runfiles discovery requires having this information available at a known
location next to the executable.

This commit emits the executable target's canonical repository name as
the first line of the repository mapping manifest.