[Skymeld] Use a separate event for TopLevelTargetReadyForSymlinkPlanting.
The previous setup (using TopLevelTargetAnalyzedEvent with an optionally filled
transitivePackagesForSymlinkPlanting) cause a subtle bug with the
BuildProgressPublisher with --discard_analysis_cache:

1. TopLevelTargetAnalyzedEvent(//foo, <pkgsForSymlinkPlanting>): status: //foo:
ANALYSIS -> TEST_FILTERING -> NOT_BUILT
...
2. TargetCompleteEvent(//foo): status: //foo: COMPLETE
3. Analysis cache discarded
4. TopLevelTargetAnalyzedEvent(//foo, []): status: //foo: COMPLETE -> ANALYSIS
(!!! Error !!!)

The 4th step happens because the de-duplication mechanism in
SkymeldBuildProgressPublisher failed to filter out the 2nd
TopLevelStatusAnalyzedEvent: it's not equal to the first one (the packages are
already discarded by step 3).

Having a separate TopLevelTargetReadyForSymlinkPlanting event would fix this
issue. From the perspective of BuildProgressPublisher, we don't care about
whether there's symlink planting or not.

PiperOrigin-RevId: 486880521
Change-Id: Icecbb67b7d938efb2741df52701ef7b0bd4bc2e0