Fix missing DocCategory values causing symbols to show up in wrong places
If the `category` field of the `@StarlarkBuiltin` annotation is not specified, it defaults to a "top-level module" category, which is often not correct. For example, in `attr.label(...)`, `attr` is a top-level module; but most of the things currently listed as top-level modules are not; instead, they should be listed as built-in types. This CL does just that.

PiperOrigin-RevId: 487211322
Change-Id: Ide6739e96e395b0bda6cd71b07e02590b0bb488c