[Skymeld] Fix a bug with action cache being re-loaded despite using --nouse_action_cache.
The method ExecutionTool#getActionCache, despite its name, may perform loading
an action cache if the actionCache is currently null. This caused the SKymeld
path to create an action cache instance despite the build having
--nouse_action_cache. This was caught by one of the integration tests.

This CL fixes that bug by having a separate method that just returns the
reference to the current in-use action cache instance, without attempting to
load it.

PiperOrigin-RevId: 487468119
Change-Id: I32a83236271cd96413f4ec54b8fa51f97ad1d828