Fix timeout and enormous gRPC log dumped from ByteStreamUploaderTest if some other tests failed.
The root cause is some retriers in the test always retry the gRPC calls no matter what the exception is. Sometimes, the exception is CANCELLED due to the test case was interrupted by the framework since other tests failed. This is effectively sending the gRPC requests in an endless loop until test timed out.

PiperOrigin-RevId: 487491332
Change-Id: Iea3d345bf581c8a24678f45151cd80b77490d53a