[Skymeld] Fix an issue where non-test targets were reported as tests when there's no target.
Assume a situation where we have a build without any target.
In regular blaze, we expect the output to be "Found 0 targets", while in Skymeld it was "Found 0 test targets".

This happened because regular blaze checks for the nullness of `analysisResult.getTargetsToTest()` to confirm that we shouldn't print "Found 0 test targets". Skymeld, on the other hand, uses the BuildResultListener which initializes targetsToTest as an empty set instead of null. This resulted in it printing "Found 0 test targets".

This CL fixes that by detaching the logic of whether or not we should report the tests from both BuildResultListener and AnalysisResult, and instead getting this info from the BuildRequest.

PiperOrigin-RevId: 487810344
Change-Id: I59e900699ab226219488ca22ae93fedbc3763d21