Fix Bash rlocation failure with stricter Bash options
When run in the context of the bazel-skylib unit test setup, the new
repo mapping logic in the Bash runfiles library failed due to a
non-matching grep in a pipe. Fix this by using a wrapper around grep
throughout that does not exit with exit code 1 if there is no match.