Fix the link for the `--test_sharding_strategy` flag in the Bazel Test encyclopedia.
The link is currently pointing to the "Running executables" section of the Bazel User Manual, but that seems unwise to me. That section is documenting the `bazel run` command, to which the `--test_sharding_strategy` flag is completely irrelevant.

Instead, I changed the link to point to the section of the Bazel Command-Line Reference that contains the helpstring for the flag.

PiperOrigin-RevId: 489209646
Change-Id: Ied40d4b3e46f73294e9e9a9088e046013edfb0b9