Limit the number of input root Merkle trees we keep in memory
Bazel currently keeps the input root Merkle tree in memory for the
entire lifetime of the Spawn. It does this, because it needs this data
for two purposes:

- To compute the Action digest, so that we can do lookups against the
  remote cache.
- To upload it into the CAS in case of a remote cache miss.

The latter may be retried in case of transient RPC failures, as
controlled by the --remote_retries flag.

The downside of this approach is that if Bazel is invoked with a high
number of --jobs, we can easily exhaust memory. This is why we make the
following two changes:

- When computing the Action digest, we now acquire/release a semaphore,
  limiting the concurrency to the number of available CPU cores. This
  ought to be sufficient, as computing Merkle trees is CPU bound anyway.

- We change RemoteAction to no longer store the Merkle tree. It is thus
  disposed after the Action digest is computed. This means that if we
  need to upload it into the CAS in case of a cache miss, we recompute
  it once more, merely to complete the upload phase.

This should reduce the number of Merkle trees stored in memory
significantly. Furthermore, by acquiring the semaphore in such a way
that interruptions are permitted, Bazel is able to terminate far more
quickly (in case of build failures, or when pressing ^C).

Change-Id: I72d19beb7d3a2675d3bd89b7ba25166188726d9b
Issue: BEN-14271
Reviewed-on: https://gerrit.sg.apple.com/c/swplatform/bazelbuild/bazel/+/387958
Reviewed-by: Alan Falloon <afalloon@apple.com>
Reviewed-by: Mathieu Olivari <molivari@apple.com>
Tested-by: Hyperion Jenkins <hyperion_jenkins@apple.com>